refactor: Update file naming convention
