--- src/features/funding/dto/funding-update.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { FundingCreateDto } from './funding-create.dto';
-
-export class FundingUpdateDto extends PartialType(FundingCreateDto) {}
