+++ src/features/funding/dto/update-funding.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateFundingDto } from './create-funding.dto';
+
+export class UpdateFundingDto extends PartialType(CreateFundingDto) {}
